WIP: moving editor related things into its own assembly
