+++ Packages/com.unity.inputsystem/InputSystem/Actions/Composites/AxisComposite.cs
@@ -3,12 +3,7 @@ using UnityEngine.InputSystem.Layouts;
 using UnityEngine.InputSystem.Processors;
 using UnityEngine.InputSystem.Utilities;
 
-#if UNITY_EDITOR
-using System;
-using UnityEditor;
-using UnityEngine.InputSystem.Editor;
-using UnityEngine.UIElements;
-#endif
+
 
 namespace UnityEngine.InputSystem.Composites
 {
@@ -209,40 +204,5 @@ namespace UnityEngine.InputSystem.Composites
         }
     }
 
-    #if UNITY_EDITOR
-    internal class AxisCompositeEditor : InputParameterEditor<AxisComposite>
-    {
-        private GUIContent m_WhichAxisWinsLabel = new GUIContent("Which Side Wins",
-            "Determine which axis 'wins' if both are actuated at the same time. "
-            + "If 'Neither' is selected, the result is 0 (or, more precisely, "
-            + "the midpoint between minValue and maxValue).");
-
-        public override void OnGUI()
-        {
-#if UNITY_INPUT_SYSTEM_PROJECT_WIDE_ACTIONS
-            if (!InputSystem.settings.useIMGUIEditorForAssets) return;
-#endif
-            target.whichSideWins = (AxisComposite.WhichSideWins)EditorGUILayout.EnumPopup(m_WhichAxisWinsLabel, target.whichSideWins);
-        }
-
-#if UNITY_INPUT_SYSTEM_PROJECT_WIDE_ACTIONS
-        public override void OnDrawVisualElements(VisualElement root, Action onChangedCallback)
-        {
-            var modeField = new EnumField(m_WhichAxisWinsLabel.text, target.whichSideWins)
-            {
-                tooltip = m_WhichAxisWinsLabel.tooltip
-            };
-
-            modeField.RegisterValueChangedCallback(evt =>
-            {
-                target.whichSideWins = (AxisComposite.WhichSideWins)evt.newValue;
-                onChangedCallback();
-            });
-
-            root.Add(modeField);
-        }
-
-#endif
-    }
-    #endif
+    
 }
