+++ Packages/com.unity.inputsystem/InputSystem/Actions/Interactions/TapInteraction.cs
@@ -2,10 +2,7 @@ using System;
 using System.ComponentModel;
 using UnityEngine.InputSystem.Controls;
 using UnityEngine.Scripting;
-#if UNITY_EDITOR
-using UnityEngine.InputSystem.Editor;
-using UnityEngine.UIElements;
-#endif
+
 
 namespace UnityEngine.InputSystem.Interactions
 {
@@ -94,44 +91,4 @@ namespace UnityEngine.InputSystem.Interactions
         }
     }
 
-    #if UNITY_EDITOR
-    internal class TapInteractionEditor : InputParameterEditor<TapInteraction>
-    {
-        protected override void OnEnable()
-        {
-            m_DurationSetting.Initialize("Max Tap Duration",
-                "Time (in seconds) within with a control has to be released again for it to register as a tap. If the control is held "
-                + "for longer than this time, the tap is canceled.",
-                "Default Tap Time",
-                () => target.duration, x => target.duration = x, () => InputSystem.settings.defaultTapTime);
-            m_PressPointSetting.Initialize("Press Point",
-                "The amount of actuation a control requires before being considered pressed. If not set, default to "
-                + "'Default Button Press Point' in the global input settings.",
-                "Default Button Press Point",
-                () => target.pressPoint, v => target.pressPoint = v,
-                () => InputSystem.settings.defaultButtonPressPoint);
-        }
-
-        public override void OnGUI()
-        {
-#if UNITY_INPUT_SYSTEM_PROJECT_WIDE_ACTIONS
-            if (!InputSystem.settings.useIMGUIEditorForAssets) return;
-#endif
-            m_DurationSetting.OnGUI();
-            m_PressPointSetting.OnGUI();
-        }
-
-#if UNITY_INPUT_SYSTEM_PROJECT_WIDE_ACTIONS
-        public override void OnDrawVisualElements(VisualElement root, Action onChangedCallback)
-        {
-            m_DurationSetting.OnDrawVisualElements(root, onChangedCallback);
-            m_PressPointSetting.OnDrawVisualElements(root, onChangedCallback);
-        }
-
-#endif
-
-        private CustomOrDefaultSetting m_DurationSetting;
-        private CustomOrDefaultSetting m_PressPointSetting;
-    }
-    #endif
 }
