+++ Packages/com.unity.inputsystem/InputSystem/Editor/Actions.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 51c96b8e65b5a46b28c06e3fbc751c97
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
