+++ Packages/com.unity.inputsystem/InputSystem/Editor/Actions/AxisCompositeEditor.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: f04048c85eab54bcc9c4cba980634c7c
