+++ Packages/com.unity.inputsystem/InputSystem/Editor/Actions/Interactions.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 79f4019b5557b4b8a87a46a5869d9666
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
