+++ Packages/com.unity.inputsystem/InputSystem/Editor/Actions/Interactions/SlowTapInteractionEditor.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 039c745f633954ebb99500cc5f3169b6
