+++ Packages/com.unity.inputsystem/InputSystem/Editor/Actions/Vector2CompositeEditor.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: bc26b75e5423c40078082d8e558222d0
