+++ Packages/com.unity.inputsystem/InputSystem/Editor/Actions/Vector3CompositeEditor.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 30d6e70def27c4f0bb92b90ed5d7431b
