+++ Packages/com.unity.inputsystem/InputSystem/Editor/Controls.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 51baa5495e14f4589bdd9593c6a72ae3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
