+++ Packages/com.unity.inputsystem/InputSystem/Editor/Controls/Processors.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b4690836c3a9e42f18b8d14b10b92097
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
