+++ Packages/com.unity.inputsystem/InputSystem/Editor/Controls/Processors/StickDeadzoneProcessorEditor.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 59bbf7f4fa3af4e0eb0ed7eaabff05db
