+++ Packages/com.unity.inputsystem/InputSystem/Editor/Plugins.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b873975c6661d4c3a9442c4134bb5c63
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
