+++ Packages/com.unity.inputsystem/InputSystem/Editor/Plugins/EnhancedTouch.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 802b0375540bb4a04be19fe17d88d659
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
