+++ Packages/com.unity.inputsystem/InputSystem/Editor/Plugins/EnhancedTouch/TouchSimulationEditor.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 848690e04e53948a5b5ffb210a4fa64f
