+++ Packages/com.unity.inputsystem/InputSystem/Editor/Plugins/HID.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2a5747283a8db45e99cf4536241aeb9a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
