+++ Packages/com.unity.inputsystem/InputSystem/Editor/Plugins/PlayerInput.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3797c94c0360c4497af338652e02249a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
