+++ Packages/com.unity.inputsystem/InputSystem/Editor/UnityEditor.InputSystem.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 43c9ed31071e54e6ab323e5e2f3bdc9c
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
